refactor(bot_run): unify timeout as float with single resolution point
- Resolve timeout once at BotRunner entry: metadata['timeout'] or
  config.bot_service.request_timeout (default 30.0), guaranteed non-None
- Change timeout type from int|None to float across the entire call chain
  (Protocol, dispatcher, BotService, AsyncChatClient, TaskConcurrencyPool)
  to support sub-second precision (e.g. 0.2s)
- Remove per-layer request_timeout fallback in ClawBotService and
  BaasBotService — they now transparently pass through the resolved value
- Fix double wait_for: slot.run uses original timeout as outer guard,
  bot_service.send_message receives timeout - 0.2 (floored at 0.1) as
  inner timeout, giving DB write a 200ms margin before the outer guard
  fires
- Inject default_request_timeout into BotRunner via DI container
- Add unit tests covering timeout reduction, floor edge case, and
  pool-path passthrough

+++ src/baas/src/secbaas/community/core/service/bot_run/_async_chat_client.py
@@ -271,7 +271,7 @@ class AsyncChatClient:
         message: str,
         session_key: str | None = None,
         wait_result: bool = True,
-        timeout: int | None = None,  # noqa: ASYNC109
+        timeout: float | None = None,  # noqa: ASYNC109
         auth_token: str | None = None,
         app_id: str | None = None,
         chat_metadata: dict[str, str] | None = None,
@@ -412,7 +412,7 @@ class AsyncChatClient:
         self,
         message: str,
         session_key: str | None = None,
-        timeout: int | None = None,  # noqa: ASYNC109
+        timeout: float | None = None,  # noqa: ASYNC109
         auth_token: str | None = None,
         app_id: str | None = None,
         chat_metadata: dict[str, str] | None = None,
@@ -576,7 +576,7 @@ class AsyncChatClient:
     @staticmethod
     async def _drain_stream_queue(
         queue: asyncio.Queue[StreamChunk],
-        timeout: int | None,
+        timeout: float | None,
     ) -> AsyncIterator[StreamChunk]:
         """从 stream_queue 消费 StreamChunk，遇到终止 chunk 后停止。
